feature: Check size before upload
Files were uploaded and then the size was checked.
Javascript can now check the size of the files before they are
uploaded. If the total size of the files exceeds 40MB an alert is
raised. The "Share files button" is disabled on page load. If
total file size is less than 40MB the button is enabled.